+++ thunderstore/CHANGELOG.md
@@ -1,3 +1,11 @@
 > # Update Information (Latest listed first)
 > ### v0.1.0
 > - fist version
+> ### v0.2.0
+> - bug fixes
+> - minor refactor
+> - added settings to modify bosses' hp and regen
+> - added skill curve settings
+> - added changer for enemy level chances (stars)
+> - added possibility to change death penalty for skills
+> - added toggle to allow building in cleared dungeons